+++ src/climate/checktitle.c
@@ -1,10 +1,10 @@
 /**************************************************************************************/
 /**                                                                                \n**/
-/**               c  h  e  c  k  c  l  i  m  a  t  e  t  i  t  l  e  .  c          \n**/
+/**               c  h  e  c  k  t  i  t  l  e  .  c                               \n**/
 /**                                                                                \n**/
 /**     C implementation of LPJmL                                                  \n**/
 /**                                                                                \n**/
-/**     Function checks for matching titles for climate input files                \n**/
+/**     Function checks for matching titles for input files                        \n**/
 /**                                                                                \n**/
 /** (C) Potsdam Institute for Climate Impact Research (PIK), see COPYRIGHT file    \n**/
 /** authors, and contributors see AUTHORS file                                     \n**/
@@ -16,11 +16,12 @@
 
 #include "lpj.h"
 
-Bool checkclimatetitle(const Attr *attrs,    /**< pointer to array of attributes or NULL */
-                       int n_attr,           /**< size of array attribute */
-                       const char *filename, /**< filename of climate input file */
-                       Config *config        /**< LPJmL configuration */
-                      )                      /** \return FALSE on matching titles */
+Bool checktitle(const Attr *attrs,    /**< pointer to array of attributes or NULL */
+                int n_attr,           /**< size of array attribute */
+                const char *filename, /**< filename of climate input file */
+                char **title,         /**< title or NULL */
+                Bool isout            /**< output to stderr (TRUE/FALSE) */
+               )                      /** \return FALSE on matching titles */
 {
   char *climate;
   Bool rc=FALSE;
@@ -29,15 +30,15 @@ Bool checkclimatetitle(const Attr *attrs,    /**< pointer to array of attributes
     climate=getattr(attrs,n_attr,"title");
     if(climate!=NULL)
     {
-      if(config->climate==NULL)
-        config->climate=climate;
+      if(*title==NULL)
+        *title=climate;
       else
       {
-        if(strcmp(climate,config->climate))
+        if(strcmp(climate,*title))
         {
-          if(isroot(*config))
-            fprintf(stderr,"WARNING043: Climate %s in file '%s' differs from %s.\n",
-                    climate,filename,config->climate);
+          if(isout)
+            fprintf(stderr,"WARNING043: Title %s in file '%s' differs from %s.\n",
+                    climate,filename,*title);
           rc=TRUE;
         }
         free(climate);
@@ -45,4 +46,4 @@ Bool checkclimatetitle(const Attr *attrs,    /**< pointer to array of attributes
     }
   }
   return rc;
-} /* of 'checkclimatetitle' */
+} /* of 'checktitle' */