+++ src/lpj/fscanconfig.c
@@ -228,6 +228,7 @@ Bool fscanconfig(Config *config,    /**< LPJ configuration */
   config->coupler_in=0;
   config->socket=NULL;
   config->climate=NULL;
+  config->landuse=NULL;
   if (verbose>=VERB) puts("// I. type section");
   config->coupled_model=NULL;
   if(iskeydefined(file,"coupled_model"))