+++ CHANGELOG.md
@@ -29,7 +29,7 @@ of `major.minor.patch` with
  - Base temp and hlimit values are read from crop PHU JSON or NetCDF file and compared to values read from the PFT JSON file. `lpjml` stops in case of different values.
  - Warning is printed if no JSON or NetCDF file is read for crop PHU input or file contains no `"basetemp"` or `"hlimit"` data.
  - Band names for mapping are read from JSON metafiles and NetCDF input files.
- - Globale attribute `"title"` is read from climate input files and checked for matching titles.
+ - Globale attribute `"title"` is read from climate and land-use input files and checked for matching titles.
  - Global attribute `"climate"` is read for crop PHU and livestock density output and compared to title of climate input.
  - File `magic` added to support LPJmL clm and restart files for the `file` command. Content of the magic file has to copied to the `~/.magic` file.